reorganize objects into their own package
