+++ kubetest/objects/__init__.py
@@ -1,0 +1,12 @@
+"""Kubetest wrappers around Kubernetes API Objects."""
+
+# flake8: noqa
+
+from .api_object import ApiObject
+from .clusterrolebinding import ClusterRoleBinding
+from .configmap import ConfigMap
+from .container import Container
+from .deployment import Deployment
+from .pod import Pod
+from .secret import Secret
+from .service import Service
